refactor(asset): decompose AssetService into Lifecycle/Write/Query + cost-math util
Behavior-preserving regulated money-path (5 JE-posting $transaction). Facade keeps
22-method surface + 6-arg ctor + delegates; internally constructs 3 sub-services
(spec new-site + module untouched). 7 $tx moved WHOLE (5 JE-posting incl
markInvoiceReceived TOCTOU rollback; V15 period guards + BLOCKED audit preserved);
cost math (VAT-inclusive + WHT legality) → pure util. tsc 0, asset tests green, lint 0.
